change quantity type int to long

+++ app/src/main/java/uz/behzod/eightytwenty/utils/constants/Water.kt
@@ -2,7 +2,7 @@ package uz.behzod.eightytwenty.utils.constants
 
 import uz.behzod.eightytwenty.R
 
-enum class Water(val quantity: Int, val image: Int) {
+enum class Water(val quantity: Long, val image: Int) {
     HOT_CUP(200,R.drawable.hot_cup),
     MUG(400, R.drawable.mug),
     BOTTLE(600,R.drawable.water_bottle),
